Finished adding screen elements, created skeleton for Recycler View adapter & viewholder, re-organized classes
