+++ .idea/misc.xml
@@ -6,6 +6,7 @@
         <entry key="app/src/main/res/layout/account_fragment.xml" value="0.13267326732673268" />
         <entry key="app/src/main/res/layout/main_activity.xml" value="0.19610507246376813" />
         <entry key="app/src/main/res/layout/main_fragment.xml" value="0.19610507246376813" />
+        <entry key="app/src/main/res/layout/transaction_row.xml" value="0.15260416666666668" />
         <entry key="app/src/main/res/menu/bottom_nav_menu.xml" value="0.2864583333333333" />
       </map>
     </option>
