+++ app/src/main/java/com/akul/capwayinterview/models/TransactionModel.kt
@@ -1,0 +1,11 @@
+package com.akul.capwayinterview.models
+
+import android.media.Image
+import java.util.Date
+
+data class TransactionModel(
+    val vendorName: String,
+    val dateTime: Date,
+    val icon: Image,
+    val amount: String
+)
