+++ app/src/main/java/com/akul/capwayinterview/ui/fragments/AccountFragment.kt
@@ -1,6 +1,5 @@
-package com.akul.capwayinterview.ui.main
+package com.akul.capwayinterview.ui.fragments
 
-import android.accounts.Account
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,9 +7,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.fragment.findNavController
-import com.akul.capwayinterview.R
+import com.akul.capwayinterview.models.MainViewModel
 import com.akul.capwayinterview.databinding.AccountFragmentBinding
 
 class AccountFragment : Fragment() {
