+++ app/src/main/java/com/akul/capwayinterview/ui/fragments/PhundsFragment.kt
@@ -1,4 +1,4 @@
-package com.akul.capwayinterview.ui.main
+package com.akul.capwayinterview.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -13,10 +13,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [SendFragment.newInstance] factory method to
+ * Use the [PhundsFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class SendFragment : Fragment() {
+class PhundsFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +34,7 @@ class SendFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_send, container, false)
+        return inflater.inflate(R.layout.fragment_phunds, container, false)
     }
 
     companion object {
@@ -44,12 +44,12 @@ class SendFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment SendFragment.
+         * @return A new instance of fragment PhundsFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            SendFragment().apply {
+            PhundsFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
