+++ app/src/main/res/layout/main_activity.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity" >
+    tools:context=".ui.MainActivity" >
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host_fragment"
