+++ app/src/main/res/values/strings.xml
@@ -18,4 +18,10 @@
     <string name="card_label">CapWay Card</string>
     <string name="card_active_label">Card 2881 is inactive</string>
     <string name="pending_charge_example">$0.00</string>
+    <string name="virtual_card_button">Virtual Card</string>
+    <string name="add_funds">Add Funds</string>
+    <string name="middle_tab_view_transactions">Transactions</string>
+    <string name="middle_tab_view_deposits">Deposits</string>
+    <string name="middle_tab_view_withdrawals">Withdrawals</string>
+    <string name="search_hint">Search</string>
 </resources>
